Refactor commands
* Move core commands from commands.py to seperate module

* Make geoip into a script
